+++ libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: sangyepa <sangyepa@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/16 22:18:39 by sangyepa          #+#    #+#             */
-/*   Updated: 2023/03/26 01:29:46 by sangyepa         ###   ########.fr       */
+/*   Updated: 2023/04/09 20:23:22 by sangyepa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
