+++ libft/ft_strmapi.c
@@ -6,7 +6,7 @@
 /*   By: sangyepa <sangyepa@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/29 23:22:49 by sangyepa          #+#    #+#             */
-/*   Updated: 2023/03/31 10:30:08 by sangyepa         ###   ########.fr       */
+/*   Updated: 2023/04/09 20:24:10 by sangyepa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,8 +16,8 @@ char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
 	char	*tmp;
 	char	*result;
-	int		len;
-	int		i;
+	size_t	len;
+	size_t	i;
 
 	tmp = (char *)s;
 	len = ft_strlen((char *)s);
@@ -27,7 +27,7 @@ char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 	i = 0;
 	while (i < len)
 	{
-		result[i] = f(i, tmp[i]);
+		result[i] = f((unsigned int)i, tmp[i]);
 		i++;
 	}
 	result[i] = 0;
